+++ lib/rails_wayback.rb
@@ -7,6 +7,7 @@ require "rails_wayback/render_provenance"
 require "rails_wayback/render_tracker"
 require "rails_wayback/toggle"
 require "rails_wayback/git"
+require "rails_wayback/ref_policy"
 require "rails_wayback/view_source"
 require "rails_wayback/doctor"